+++ src/lib/how/cards/sz-how-card-base.component.scss
