+++ src/lib/how/cards/sz-how-final-entity-card.component.html
