+++ src/lib/how/cards/sz-how-virtual-entity-card.component.html
