+++ src/lib/how/cards/sz-how-virtual-entity-card.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/how-ec-card';
+@import '../../scss/how-ec-card';
 
 :host {
     .steps-header {
