+++ src/lib/how/sz-how-entity.component.html
