+++ src/lib/how/sz-how-step-node.component.scss
