+++ src/lib/how/sz-how-step-stack.component.scss
