+++ src/lib/how/sz-how-virtual-entity-dialog.component.scss
