+++ src/public-api.ts
@@ -81,14 +81,14 @@ export { SzWhyEntityDialog } from './lib/why/sz-why-entity.component';
 export { SzWhyEntitiesDialog } from './lib/why/sz-why-entities.component';
 
 /** how */
-export { SzHowEntityComponent } from './lib/how/rc/sz-how-entity.component';
-export { SzHowStepNodeComponent } from './lib/how/rc/sz-how-step-node.component'
-export { SzHowFinalEntityCardComponent } from './lib/how/rc/cards/sz-how-final-entity-card.component';
-export { SzHowStepCardComponent } from './lib/how/rc/cards/sz-how-step-card.component';
-export { SzHowStepStackComponent } from './lib/how/rc/sz-how-step-stack.component';
-export { SzHowNavComponent } from './lib/how/rc/sz-how-nav.component';
-export { SzHowVirtualEntityCardComponent } from './lib/how/rc/cards/sz-how-virtual-entity-card.component';
-export { SzHowVirtualEntityDialog } from './lib/how/rc/sz-how-virtual-entity-dialog.component';
+export { SzHowEntityComponent } from './lib/how/sz-how-entity.component';
+export { SzHowStepNodeComponent } from './lib/how/sz-how-step-node.component'
+export { SzHowFinalEntityCardComponent } from './lib/how/cards/sz-how-final-entity-card.component';
+export { SzHowStepCardComponent } from './lib/how/cards/sz-how-step-card.component';
+export { SzHowStepStackComponent } from './lib/how/sz-how-step-stack.component';
+export { SzHowNavComponent } from './lib/how/sz-how-nav.component';
+export { SzHowVirtualEntityCardComponent } from './lib/how/cards/sz-how-virtual-entity-card.component';
+export { SzHowVirtualEntityDialog } from './lib/how/sz-how-virtual-entity-dialog.component';
 
 /** export some members of rest client to ease type use */
 export {
